Django/requirements.txt
pillow 
djangorestframework 
pymysql 
mysqlclient==1.4.6
dj-rest-auth 
django-allauth 
django-cors-headers